# openapi/ward.v0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ward API",
    "version": "0.1.0",
    "description": "Ward is tenant containment for AI SaaS: an OpenAI-compatible egress\nproxy plus a tenant control plane. Route agent/tool calls through\nWard with a tenant ID header; operators constrain, pause, resume,\nand audit one tenant while the rest keep running.\n\nHTTP/OpenAPI is Ward's primary integration contract. This document\ndescribes the implemented v0 surface.\n\nHonesty notes (prototype status):\n- Control auth is a single shared bearer token (prototype), NOT\n  production RBAC. It guards mutating /ward/* routes only when\n  WARD_REQUIRE_CONTROL_TOKEN=true; read endpoints are open.\n- POST /v1/chat/completions is OpenAI-compatible enough for the\n  demo and simple integrations; it is not a full OpenAI spec clone\n  (no tool-call passthrough guarantees in mock mode, no SSE\n  pass-through support; deterministic mock streaming only).\n  dev-only prototype disabled by default; kubernetes is planned.\n- Single-node, prototype/demo-supported operation. Production\n  readiness is not claimed.\n"
  },
  "servers": [
    {
      "url": "http://localhost:4317",
      "description": "Local/self-hosted Ward"
    }
  ],
...
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "proxy"
        ],
        "operationId": "createChatCompletion",
        "summary": "OpenAI-compatible egress endpoint with per-tenant enforcement",
        "description": "Enforcement is tenant-specific and happens before the request\nis counted or forwarded. Mock upstream by default; pass-through\nwhen WARD_UPSTREAM_OPENAI_BASE_URL and OPENAI_API_KEY are set.\nIf tenant policy lookup fails while Ward is up, behavior\nfollows WARD_PROXY_FAIL_MODE (open: allow with\nx-ward-fail-open: true response header + audit event; closed:\n503).\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/WardTenantId"
          },
          {
            "$ref": "#/components/parameters/WardOperation"
          },
          {
            "$ref": "#/components/parameters/WardCorrelationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completion. If the request body has `stream: true`, the\nresponse is a deterministic mock SSE stream\n(content-type: text/event-stream). WardError responses\n(400, 423, 429, 502, 503) are always JSON.\n",
            "headers": {
              "x-ward-fail-open": {
                "description": "Present (\"true\") only when the request was allowed via degraded fail-open",
                "schema": {
                  "type": "string"
                }
              },
              "x-ward-would-block": {
                "description": "Present (\"paused\" or \"constrained\") only when the\nrequest would have been blocked in enforce mode but\nwas allowed through in observe mode.\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"id\":\"chatcmpl-mock-...\",\"object\":\"chat.completion.chunk\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\"Ward \"},\"finish_reason\":null}]}\n\ndata: [DONE]\n"
              }
            }
          },
          "400": {
            "description": "Missing x-ward-tenant-id header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "423": {
            "description": "Tenant is paused by an operator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "429": {
            "description": "Tenant is constrained by an operator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error (pass-through mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "503": {
            "description": "Policy lookup failed and WARD_PROXY_FAIL_MODE=closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ward/workflow-runs": {
      "get": {
        "tags": [
          "workflow-runs"
        ],
        "operationId": "listWorkflowRuns",
        "summary": "List workflow runs",
        "responses": {
          "200": {
            "description": "Workflow runs, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "workflowRuns"
                  ],
                  "properties": {
                    "workflowRuns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WorkflowRun"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow-runs"
        ],
        "operationId": "createWorkflowRun",
        "summary": "Create a workflow run (control auth when enabled)",
        "description": "Containment rules: paused or constrained tenant -> run is\ncreated as \"blocked\" and never executed. Runner status: mock =\nimplemented prototype; docker = dev-only prototype, disabled\nunless WARD_ENABLE_DOCKER_RUNNER=true (allowlisted images\nonly); kubernetes = planned (fails with a clear error).\n",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created run (may already be blocked)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "workflowRun"
                  ],
                  "properties": {
                    "workflowRun": {
                      "$ref": "#/components/schemas/WorkflowRun"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing tenantId or invalid runner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/workflow-runs/{runId}": {
      "get": {
        "tags": [
          "workflow-runs"
        ],
        "operationId": "getWorkflowRun",
        "summary": "Inspect one workflow run",
        "parameters": [
          {
            "$ref": "#/components/parameters/RunIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "The run",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "workflowRun"
                  ],
                  "properties": {
                    "workflowRun": {
                      "$ref": "#/components/schemas/WorkflowRun"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          }
        }
      }
    },
    "/ward/workflow-runs/{runId}/cancel": {
      "post": {
        "tags": [
          "workflow-runs"
        ],
        "operationId": "cancelWorkflowRun",
        "summary": "Cancel a queued/running workflow run (control auth when enabled)",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RunIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "The run after cancellation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "workflowRun"
                  ],
                  "properties": {
                    "workflowRun": {
                      "$ref": "#/components/schemas/WorkflowRun"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ward/reset": {
      "post": {
        "tags": [
          "audit"
        ],
        "operationId": "resetWard",
        "summary": "Reset all in-memory/persisted demo state (control auth when enabled)",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Reset confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "reset"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "controlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Minimal shared-token control auth PROTOTYPE (not production\nRBAC). Required on mutating /ward/* routes only when the API\nruns with WARD_REQUIRE_CONTROL_TOKEN=true; otherwise these\nroutes are open (the default for local demos, loudly flagged\nin /health).\n"
      }
    },
    "parameters": {
      "WardTenantId": {
        "name": "x-ward-tenant-id",
        "in": "header",
        "required": true,
        "description": "Tenant this call is attributed to. Requests without it are rejected (400).",
        "schema": {
          "type": "string"
        }
      },
      "WardOperation": {
        "name": "x-ward-operation",
        "in": "header",
        "required": false,
        "description": "Logical operation label (e.g. agent.tool_call). Defaults to ai.chat_completion.",
        "schema": {
          "type": "string"
        }
      },
      "WardCorrelationId": {
        "name": "x-ward-correlation-id",
        "in": "header",
        "required": false,
        "description": "Caller-supplied correlation id, echoed into audit evidence.",
        "schema": {
          "type": "string"
        }
      },
...
      "RunIdPath": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Control auth required and token missing/invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WardError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "TenantState": {
        "type": "string",
        "enum": [
          "running",
          "constrained",
          "paused"
        ]
      },
...
      "WorkflowRunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled",
          "blocked"
        ]
      },
      "WorkflowRun": {
        "type": "object",
        "required": [
          "id",
          "tenantId",
          "runner",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "runner": {
            "type": "string",
            "enum": [
              "mock",
              "docker",
              "kubernetes"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowRunStatus"
          },
          "image": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "result": {},
          "error": {
            "type": "string"
          }
        }
      },
      "WorkflowRunRequest": {
        "type": "object",
        "required": [
          "tenantId",
          "runner"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "runner": {
            "type": "string",
            "enum": [
              "mock",
              "docker",
              "kubernetes"
            ]
          },
          "image": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WardError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. ward_tenant_constrained, ward_control_unauthorized."
          },
          "message": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TenantState"
          }
        }
      },
...
      "ChatCompletionRequest": {
        "type": "object",
        "description": "OpenAI-compatible enough for the demo; not a full spec clone.",
        "properties": {
          "model": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                }
              }
            }
          },
          "stream": {
            "type": "boolean",
            "default": false,
            "description": "If true, Ward returns a deterministic mock SSE stream\n(text/event-stream). Streaming is mock-upstream only;\npass-through streaming is not yet supported. WardError\nresponses (400, 423, 429, 502, 503) are always JSON.\n"
          }
        },
        "additionalProperties": true
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "created": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "ward": {
            "type": "object",
            "description": "Present in mock mode; identifies tenant attribution and mock status.",
            "properties": {
              "mode": {
                "type": "string"
              },
              "tenantId": {
                "type": "string"
              },
              "note": {
                "type": "string"
              }
            }
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer"
                },
                "message": {
                  "type": "object",
                  "properties": {
                    "role": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    }
                  }
                },
                "finish_reason": {
                  "type": "string"
                }
              }
            }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": {
                "type": "integer"
              },
              "completion_tokens": {
                "type": "integer"
              },
              "total_tokens": {
                "type": "integer"
              }
            }
          }
        },
        "additionalProperties": true
      }
    }
  }
}